put click and change handle functions to dedicated files
In order to separate presentational and business logic, handle change or handle click helper functions should remain in dedicated files and should not be part of App.js
